splist sso/cache.go into multiple files
- cache.go was very large- break up into 4 different files based on
  function
- move profile function from settings to profiles.go
- improve cache unit test coverage
