+++ internal/sso/cache.go
@@ -19,24 +19,16 @@ package sso
  */
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"os"
-	"slices"
-	"strconv"
-	"strings"
 	"time"
 
-	// "github.com/davecgh/go-spew/spew"
-	"github.com/synfinatic/aws-sso-cli/internal/awsparse"
 	"github.com/synfinatic/aws-sso-cli/internal/fileutils"
-	"github.com/synfinatic/aws-sso-cli/internal/tags"
 )
 
 const (
-	CACHE_VERSION      = 4
-	SLOW_FETCH_SECONDS = 2 // number of seconds before notifying users
+	CACHE_VERSION = 4
 )
 
 type SSOCache struct {
@@ -177,244 +169,6 @@ func (c *SSOCache) NeedsRefresh(s *SSOConfig, settings *Settings) bool {
 	return checkHash != c.ConfigHash
 }
 
-// AddHistory adds a role ARN to the History list up to the max number of entries
-// and then removes the History tag from any roles that aren't in our list
-func (c *Cache) AddHistory(item string) {
-	// If it's already in the list, remove item
-	c.deleteHistoryItem(item)
-
-	c.GetSSO().History = append([]string{item}, c.GetSSO().History...) // push on top
-	for int64(len(c.GetSSO().History)) > c.settings.HistoryLimit {
-		// remove the oldest entry
-		c.GetSSO().History = c.GetSSO().History[:len(c.GetSSO().History)-1]
-	}
-
-	// Update our Tags for this new item
-	aId, roleName, _ := awsparse.ParseRoleARN(item)
-	if a, ok := c.GetSSO().Roles.Accounts[aId]; ok {
-		if r, ok := a.Roles[roleName]; ok {
-			r.Tags["History"] = fmt.Sprintf("%s:%s,%d", a.Alias, roleName, time.Now().Unix())
-		}
-	}
-
-	// remove any history tags not in our list
-	roles := c.GetSSO().Roles.MatchingRolesWithTagKey("History")
-
-	for _, role := range roles {
-		exists := false
-		for _, history := range c.GetSSO().History {
-			if history == (*role).Arn {
-				exists = true
-				break
-			}
-		}
-
-		// remove any History tag for roles which don't exist in c.GetSSO().History
-		if !exists {
-			aId, roleName, _ := awsparse.ParseRoleARN(role.Arn)
-			delete(c.GetSSO().Roles.Accounts[aId].Roles[roleName].Tags, "History")
-		}
-	}
-}
-
-func (c *Cache) deleteHistoryItem(arn string) {
-	for i, value := range c.GetSSO().History {
-		if arn == value {
-			c.GetSSO().History = append(c.GetSSO().History[:i], c.GetSSO().History[i+1:]...)
-			break
-		}
-	}
-}
-
-// deleteOldHistory removes any items from history which are older than HistoryMinutes
-// Does not actually save to disk, only updates in memory cache
-func (c *Cache) deleteOldHistory() {
-	if c.settings.HistoryMinutes <= 0 {
-		// no op if HistoryMinutes <= 0
-		return
-	}
-
-	cache := c.GetSSO()
-
-	newHistoryItems := []string{}
-
-	// iteratate over each ARN in our History list
-	for _, arn := range cache.History {
-		id, role, err := awsparse.ParseRoleARN(arn)
-		if err != nil {
-			log.Debug("Unable to parse History ARN", "arn", arn, "error", err.Error())
-			c.deleteHistoryItem(arn)
-			continue
-		}
-
-		// for the given ARN, lookup the History tag
-		if a, ok := cache.Roles.Accounts[id]; ok {
-			if r, ok := a.Roles[role]; ok {
-				// figure out if this history item has expired
-				history, ok := r.Tags["History"]
-				if !ok || history == "" {
-					// doesn't have anything to expires
-					log.Debug("ARN in history list without a History tag in cache?", "arn", arn)
-					c.deleteHistoryItem(arn)
-					continue
-				}
-
-				values := strings.SplitN(history, ",", 2)
-				if len(values) != 2 {
-					log.Debug("Too few fields for History Tag", "arn", r.Arn, "history", history)
-					c.deleteHistoryItem(arn)
-					continue
-				}
-				lastTime, err := strconv.ParseInt(values[1], 10, 64)
-				if err != nil {
-					log.Debug("Unable to parse History Tag", "arn", r.Arn, "history", history, "error", err.Error())
-					c.deleteHistoryItem(arn)
-					continue
-				}
-
-				d := time.Since(time.Unix(lastTime, 0))
-				if int64(d.Minutes()) < c.settings.HistoryMinutes {
-					// keep current entries in our list
-					newHistoryItems = append(newHistoryItems, arn)
-				} else {
-					// else, delete the tag and remove the item from History by
-					// not appending it to newHistoryItems
-					delete(r.Tags, "History")
-					c.deleteHistoryItem(arn)
-					log.Debug("Removed expired history role", "arn", r.Arn)
-				}
-			} else {
-				c.deleteHistoryItem(arn)
-				log.Debug("History contains but no role by that name", "arn", arn)
-			}
-		} else {
-			c.deleteHistoryItem(arn)
-			log.Debug("History contains but no account by that name", "arn", arn)
-		}
-	}
-
-	c.GetSSO().History = newHistoryItems
-}
-
-// Refresh updates our cached Roles based on AWS SSO & our Config
-// but does not save this data!  Returns the number of roles added/deleted
-func (c *Cache) Refresh(sso *AWSSSO, config *SSOConfig, ssoName string, threads int) (int, int, error) {
-	// Only refresh once per execution
-	if c.refreshed {
-		return 0, 0, nil
-	}
-	c.refreshed = true
-	log.Debug("refreshing SSO cache", "SSOname", ssoName)
-
-	// save role creds expires time
-	expires := map[string]int64{}
-	cache := c.GetSSO()
-	now := time.Now().Unix()
-	for _, account := range cache.Roles.Accounts {
-		for _, role := range account.Roles {
-			if role.Expires > now {
-				expires[role.Arn] = role.Expires
-			}
-		}
-	}
-
-	// save existing History tags
-	historyTags := map[string]string{}
-	for _, arn := range c.SSO[ssoName].History {
-		roleFlat, err := c.GetRole(arn)
-		if err != nil {
-			continue
-		}
-		if value, ok := roleFlat.Tags["History"]; ok {
-			historyTags[arn] = value
-		}
-	}
-
-	// zero out our current roles cache entries so they don't get merged
-	oldRoles := cache.Roles.GetAllRoles()
-	oldRoleArns := []string{}
-	for _, role := range oldRoles {
-		oldRoleArns = append(oldRoleArns, role.Arn)
-	}
-
-	c.SSO[ssoName].Roles = &Roles{}
-	c.SSO[ssoName].ConfigHash = config.GetConfigHash(c.settings.ProfileFormat)
-
-	// load our AWSSSO & Config
-	r, err := c.NewRoles(sso, config, threads)
-	if err != nil {
-		return 0, 0, err
-	}
-	c.SSO[ssoName].Roles = r
-
-	// figure out what roles were added/deleted
-	newRoles := c.SSO[ssoName].Roles.GetAllRoles()
-	newRoleArns := []string{}
-	for _, role := range newRoles {
-		newRoleArns = append(newRoleArns, role.Arn)
-	}
-
-	// do we have any new or deleted roles?
-	added, deleted := 0, 0
-	for _, arn := range newRoleArns {
-		if !slices.Contains(oldRoleArns, arn) {
-			added++
-		}
-	}
-	for _, arn := range oldRoleArns {
-		if !slices.Contains(newRoleArns, arn) {
-			deleted++
-		}
-	}
-
-	// restore any configured roles that are not in AWS SSO
-	for aId, account := range config.Accounts {
-		accountId, err := awsparse.AccountIdToInt64(aId)
-		if err != nil {
-			return 0, 0, fmt.Errorf("unable to parse accountId from config.yaml %s: %s", aId, err.Error())
-		}
-		// if aId is scientific notation, convert to int64 string
-		aId, _ := awsparse.AccountIdToString(accountId)
-		for rName, role := range account.Roles {
-			if role.Via != "" {
-				log.Info("restoring via role", "role", rName, "account", accountId)
-				if _, ok := cache.Roles.Accounts[accountId]; !ok {
-					c.SSO[ssoName].Roles.Accounts[accountId] = &AWSAccount{}
-				}
-				if role.Tags == nil {
-					role.Tags = map[string]string{
-						"AccountAlias": c.SSO[ssoName].Roles.Accounts[accountId].Alias,
-						"AccountID":    aId,
-						"Email":        c.SSO[ssoName].Roles.Accounts[accountId].EmailAddress,
-						"Role":         rName,
-					}
-				}
-				c.SSO[ssoName].Roles.Accounts[accountId].Roles[rName] = &AWSRole{
-					Arn:           role.ARN,
-					DefaultRegion: role.DefaultRegion,
-					Profile:       role.Profile,
-					Tags:          role.Tags,
-					Via:           role.Via,
-				}
-			}
-		}
-	}
-
-	// restore our history tags & expires
-	for _, account := range c.SSO[ssoName].Roles.Accounts {
-		for _, role := range account.Roles {
-			if value, ok := historyTags[role.Arn]; ok {
-				role.Tags["History"] = value
-			}
-			if value, ok := expires[role.Arn]; ok {
-				role.Expires = value
-			}
-		}
-	}
-	c.ConfigCreatedAt = config.CreatedAt()
-	return added, deleted, nil
-}
-
 // pruneSSO removes any SSO instances that are no longer configured
 func (c *Cache) PruneSSO(settings *Settings) {
 	log.Debug("pruning our cache of outdated SSO instances")
@@ -432,291 +186,4 @@ func (c *Cache) PruneSSO(settings *Settings) {
 			delete(c.SSO, sso)
 		}
 	}
-}
-
-// SetRoleExpires updates the Expires time in the cache.  expires is Unix epoch time in sec
-func (c *Cache) SetRoleExpires(arn string, expires int64) error {
-	flat, err := c.GetRole(arn)
-	if err != nil {
-		return err
-	}
-
-	cache := c.GetSSO()
-	cache.Roles.Accounts[flat.AccountId].Roles[flat.RoleName].Expires = expires
-	return c.Save(false)
-}
-
-// MarkRolesExpired marks all IAM role credentials in the cache as expired
-func (c *Cache) MarkRolesExpired() error {
-	cache := c.GetSSO()
-	for accountId := range cache.Roles.Accounts {
-		for _, role := range cache.Roles.Accounts[accountId].Roles {
-			(*role).Expires = 0
-		}
-	}
-	return c.Save(false)
-}
-
-// returns all tags, but with with spaces replaced with underscores
-func (c *Cache) GetAllTagsSelect() *tags.TagsList {
-	cache := c.GetSSO()
-	t := cache.Roles.GetAllTags()
-	fixedTags := tags.NewTagsList()
-	for k, values := range *t {
-		key := strings.ReplaceAll(k, " ", "_")
-		for _, v := range values {
-			if key == "History" {
-				v = tags.ReformatHistory(v)
-			}
-			fixedTags.Add(key, strings.ReplaceAll(v, " ", "_"))
-		}
-	}
-	return fixedTags
-}
-
-// GetRoleTagsSelect returns all the tags for each role with all the spaces
-// replaced with underscores
-func (c *Cache) GetRoleTagsSelect() *RoleTags {
-	ret := RoleTags{}
-	cache := c.GetSSO()
-	fList := cache.Roles.GetAllRoles()
-	for _, role := range fList {
-		ret[role.Arn] = map[string]string{}
-		for k, v := range role.Tags {
-			key := strings.ReplaceAll(k, " ", "_")
-			if key == "History" {
-				v = tags.ReformatHistory(v)
-			}
-			value := strings.ReplaceAll(v, " ", "_")
-			ret[role.Arn][key] = value
-		}
-	}
-	return &ret
-}
-
-// GetRole returns the AWSRoleFlat for the given role ARN
-func (c *Cache) GetRole(arn string) (*AWSRoleFlat, error) {
-	accountId, roleName, err := awsparse.ParseRoleARN(arn)
-	if err != nil {
-		return &AWSRoleFlat{}, err
-	}
-	cache := c.GetSSO()
-	return cache.Roles.GetRole(accountId, roleName)
-}
-
-// Merges the AWS SSO and our Config file to create our Roles struct
-// which is defined in cache_roles.go
-func (c *Cache) NewRoles(as *AWSSSO, config *SSOConfig, threads int) (*Roles, error) {
-	r := Roles{
-		SSORegion:     config.SSORegion,
-		StartUrl:      config.StartUrl,
-		DefaultRegion: config.DefaultRegion,
-		Accounts:      map[int64]*AWSAccount{},
-		SSOName:       config.settings.DefaultSSO,
-	}
-
-	if err := c.addSSORoles(&r, as, threads); err != nil {
-		return &Roles{}, err
-	}
-
-	if err := c.addConfigRoles(&r, config); err != nil {
-		return &Roles{}, err
-	}
-
-	if err := r.CheckProfiles(c.settings); err != nil {
-		return &Roles{}, err
-	}
-
-	return &r, nil
-}
-
-// goroutine worker to fetch the RoleInfo for the given account
-func fetchSSORole(id int, as *AWSSSO, aInfo <-chan AccountInfo, rInfo chan<- []RoleInfo) {
-	for {
-		a := <-aInfo
-		if a.AccountId == "" {
-			// need some way to exit our worker...
-			break
-		}
-		log.Debug("Worker processing", "worker", id, "accountID", a.AccountId)
-		roles, err := as.GetRoles(a)
-		if err != nil {
-			panic(fmt.Sprintf("Unable to get AWS SSO roles: %s", err.Error()))
-		}
-		rInfo <- roles
-	}
-}
-
-// processSSORoles updates the *Roles with ith the list of RoleInfo
-// and using our SSOCache
-func processSSORoles(roles []RoleInfo, cache *SSOCache, r *Roles) {
-	for _, role := range roles {
-		log.Debug(fmt.Sprintf("Processing %s:%s", role.AccountId, role.RoleName))
-		accountId := role.GetAccountId64()
-
-		if _, ok := r.Accounts[accountId]; !ok {
-			r.Accounts[accountId] = &AWSAccount{
-				Alias:        role.AccountName, // AWS SSO calls it `AccountName`
-				EmailAddress: role.EmailAddress,
-				Tags:         map[string]string{},
-				Roles:        map[string]*AWSRole{},
-			}
-		}
-
-		r.Accounts[accountId].Roles[role.RoleName] = &AWSRole{
-			Arn: awsparse.MakeRoleARN(accountId, role.RoleName),
-			Tags: map[string]string{
-				"AccountID":    role.AccountId,
-				"AccountAlias": role.AccountName, // AWS SSO calls it `AccountName`
-				"Email":        role.EmailAddress,
-				"Role":         role.RoleName,
-			},
-		}
-		// need to copy over the Expires & History fields from our current cache
-		if _, ok := cache.Roles.Accounts[accountId]; ok {
-			if _, ok := cache.Roles.Accounts[accountId].Roles[role.RoleName]; ok {
-				if expires := cache.Roles.Accounts[accountId].Roles[role.RoleName].Expires; expires > 0 {
-					r.Accounts[accountId].Roles[role.RoleName].Expires = expires
-				}
-				if v, ok := cache.Roles.Accounts[accountId].Roles[role.RoleName].Tags["History"]; ok {
-					r.Accounts[accountId].Roles[role.RoleName].Tags["History"] = v
-				}
-			}
-		}
-	}
-}
-
-// addSSORoles retrieves all the SSO Roles from AWS SSO and places them in r
-func (c *Cache) addSSORoles(r *Roles, as *AWSSSO, threads int) error {
-	cache := c.GetSSO()
-
-	accounts, err := as.GetAccounts()
-	if err != nil {
-		return fmt.Errorf("unable to get list of AWS accounts via AWS SSO: %s", err.Error())
-	}
-
-	if len(accounts) == 0 {
-		return fmt.Errorf("no AWS accounts found in AWS SSO")
-	}
-
-	// Our first query must NOT be part of the worker pool so our AccessToken
-	// can be updated
-	firstJob, accounts := accounts[0], accounts[1:]
-	roles, err := as.GetRoles(firstJob)
-	if err != nil {
-		return err
-	}
-	processSSORoles(roles, cache, r)
-
-	// Per #448, doing this serially is too slow for many accounts.  Hence,
-	// we'll use a worker pool.
-	if len(accounts) > 0 {
-		workers := c.settings.Threads
-		if threads > 0 {
-			workers = threads
-		}
-		if workers > len(accounts) {
-			workers = len(accounts)
-		}
-
-		tasks := make(chan AccountInfo, len(accounts))
-		results := make(chan []RoleInfo, len(accounts))
-
-		// feed our workers with our other accounts
-		for _, aInfo := range accounts {
-			tasks <- aInfo
-		}
-		close(tasks)
-
-		// start our workers...
-		for w := 1; w <= workers; w++ {
-			go fetchSSORole(w, as, tasks, results)
-		}
-
-		// Notify
-		ticker := time.NewTicker(SLOW_FETCH_SECONDS * time.Second)
-
-		for count := 0; count < len(accounts); {
-			select {
-			case roles := <-results:
-				processSSORoles(roles, cache, r)
-				count++ // increment count only when processing results
-				log.Debug("proccessed", "accounts", count, "new_roles", len(roles), "total_roles", len(r.GetAllRoles()))
-			case <-ticker.C:
-				log.Warn(fmt.Sprintf("fetching roles for %d accounts, this might take a while...", len(accounts)+1))
-				ticker.Stop()
-			}
-		}
-		close(results)
-	}
-	return nil
-}
-
-type contextKey string
-
-const (
-	accountIdKey contextKey = "accountID"
-)
-
-// addConfigRoles decorates the provided Roles with the contents of our config
-func (c *Cache) addConfigRoles(r *Roles, config *SSOConfig) error {
-	// The load all the Config file stuff.  Normally this is just adding markup, but
-	// for accounts &roles that are not in SSO, we may be creating them as well!
-	for accountId, account := range config.Accounts {
-		id, err := awsparse.AccountIdToInt64(accountId)
-		if err != nil {
-			return err
-		}
-		ctx := context.WithValue(context.Background(), accountIdKey, id)
-		if _, ok := r.Accounts[id]; !ok {
-			log.DebugContext(ctx, "config.yaml defines an AWS AccountID, but you don't have access.")
-			continue
-		}
-		r.Accounts[id].DefaultRegion = account.DefaultRegion
-		r.Accounts[id].Name = account.Name
-
-		// set our account tags
-		for k, v := range config.Accounts[accountId].Tags {
-			r.Accounts[id].Tags[k] = v
-		}
-
-		// set the AWS SSO tags for all the SSO roles
-		for roleName := range r.Accounts[id].Roles {
-			aId, _ := awsparse.AccountIdToString(id)
-			r.Accounts[id].Roles[roleName].Tags["AccountID"] = aId
-			r.Accounts[id].Roles[roleName].Tags["AccountAlias"] = r.Accounts[id].Alias
-			r.Accounts[id].Roles[roleName].Tags["Email"] = r.Accounts[id].EmailAddress
-			r.Accounts[id].Roles[roleName].Tags["Role"] = roleName
-			if r.Accounts[id].Name != "" {
-				r.Accounts[id].Roles[roleName].Tags["AccountName"] = r.Accounts[id].Name
-			}
-			if r.Accounts[id].Roles[roleName].DefaultRegion != "" {
-				r.Accounts[id].Roles[roleName].Tags["DefaultRegion"] = r.Accounts[id].Roles[roleName].DefaultRegion
-			}
-		}
-
-		// set the tags from the config file
-		for roleName, role := range config.Accounts[accountId].Roles {
-			if _, ok := r.Accounts[id].Roles[roleName]; !ok {
-				log.DebugContext(ctx, "config.yaml defines a role but you don't have access", "role", roleName)
-				continue
-			}
-			r.Accounts[id].Roles[roleName].Arn = awsparse.MakeRoleARN(id, roleName)
-			r.Accounts[id].Roles[roleName].Profile = role.Profile
-			r.Accounts[id].Roles[roleName].DefaultRegion = r.Accounts[id].DefaultRegion
-			r.Accounts[id].Roles[roleName].Via = role.Via
-			if role.DefaultRegion != "" {
-				r.Accounts[id].Roles[roleName].DefaultRegion = role.DefaultRegion
-			}
-			// Copy the account tags to the role
-			for k, v := range config.Accounts[accountId].Tags {
-				r.Accounts[id].Roles[roleName].Tags[k] = v
-			}
-			// Insert role specific tags (possible overwrite of account level)
-			for k, v := range role.Tags {
-				r.Accounts[id].Roles[roleName].Tags[k] = v
-			}
-		}
-	}
-	return nil
 }
