+++ internal/sso/cache_query_test.go
@@ -1,0 +1,152 @@
+package sso
+
+/*
+ * AWS SSO CLI
+ * Copyright (c) 2021-2026 Aaron Turner  <synfinatic at gmail dot com>
+ *
+ * This program is free software: you can redistribute it
+ * and/or modify it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or with the authors permission any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import (
+	"github.com/stretchr/testify/assert"
+)
+
+func (suite *CacheTestSuite) TestGetRole() {
+	t := suite.T()
+	r, _ := suite.cache.GetRole(TEST_ROLE_ARN)
+	assert.Equal(t, int64(707513610766), r.AccountId)
+	assert.Equal(t, "AWSAdministratorAccess", r.RoleName)
+	assert.Equal(t, TEST_ROLE_ARN, r.Arn)
+	assert.Equal(t, "Dev Account", r.AccountName)
+	assert.Equal(t, "control-tower-dev-sub1-aws", r.AccountAlias)
+	assert.Equal(t, "test+control-tower-sub1@ourcompany.com", r.EmailAddress)
+	assert.Equal(t, "", r.Profile)
+	assert.Equal(t, "", r.DefaultRegion)
+	assert.Equal(t, "us-east-1", r.SSORegion)
+	assert.Equal(t, "https://d-754545454.awsapps.com/start", r.StartUrl)
+	assert.Equal(t, "Default", r.SSO)
+
+	tags := map[string]string{
+		"AccountAlias": "control-tower-dev-sub1-aws",
+		"AccountID":    "707513610766",
+		"AccountName":  "Dev Account",
+		"Email":        "test+control-tower-sub1@ourcompany.com",
+		"Role":         "AWSAdministratorAccess",
+		"Type":         "Sub Account",
+	}
+	assert.Equal(t, tags, r.Tags)
+
+	_, err := suite.cache.GetRole("arn:aws:2344:role/missing-colon")
+	assert.Error(t, err)
+}
+
+func (suite *CacheTestSuite) TestAccountIds() {
+	t := suite.T()
+
+	ids := suite.cache.GetSSO().Roles.AccountIds()
+	assert.Equal(t, 4, len(ids))
+}
+
+func (suite *CacheTestSuite) TestGetAllRoles() {
+	t := suite.T()
+
+	cache := suite.cache.GetSSO()
+	roles := cache.Roles.GetAllRoles()
+	assert.Equal(t, 19, len(roles))
+
+	aroles := cache.Roles.GetAccountRoles(707513610766)
+	assert.Equal(t, 4, len(aroles))
+	aroles = cache.Roles.GetAccountRoles(502470824893)
+	assert.Equal(t, 4, len(aroles))
+	aroles = cache.Roles.GetAccountRoles(25823461518)
+	assert.Equal(t, 7, len(aroles))
+}
+
+func (suite *CacheTestSuite) TestGetAllTags() {
+	t := suite.T()
+	cache := suite.cache.GetSSO()
+
+	tl := cache.Roles.GetAllTags()
+	assert.Equal(t, 10, len(*tl))
+	tl = suite.cache.GetAllTagsSelect()
+	assert.Equal(t, 10, len(*tl))
+}
+
+func (suite *CacheTestSuite) TestGetRoleTags() {
+	t := suite.T()
+	cache := suite.cache.GetSSO()
+
+	rt := cache.Roles.GetRoleTags()
+	assert.Equal(t, 19, len(*rt))
+	rt = suite.cache.GetRoleTagsSelect()
+	assert.Equal(t, 19, len(*rt))
+}
+
+func (suite *CacheTestSuite) TestMatchingRoles() {
+	t := suite.T()
+	cache := suite.cache.GetSSO()
+
+	match := map[string]string{
+		"Type": "Main Account",
+		"Foo":  "Bar",
+	}
+	roles := cache.Roles.MatchingRoles(match)
+	assert.Equal(t, 1, len(roles))
+
+	match["Nothing"] = "Matches"
+	roles = cache.Roles.MatchingRoles(match)
+	assert.Equal(t, 0, len(roles))
+}
+
+func (suite *CacheTestSuite) TestIsExpired() {
+	t := suite.T()
+
+	r, _ := suite.cache.GetRole(TEST_ROLE_ARN)
+	assert.True(t, r.IsExpired())
+}
+
+func (suite *CacheTestSuite) BadRole() {
+	t := suite.T()
+	_, err := suite.cache.GetRole(INVALID_ROLE_ARN)
+	assert.Error(t, err)
+
+	_, err = suite.cache.GetRole(INVALID_ACCOUNT_ARN)
+	assert.Error(t, err)
+}
+
+func (suite *CacheTestSuite) TestMarkRolesExpired() {
+	t := suite.T()
+	err := suite.cache.MarkRolesExpired()
+	assert.NoError(t, err)
+
+	sso := suite.cache.GetSSO()
+	for _, account := range sso.Roles.Accounts {
+		for _, role := range account.Roles {
+			assert.Equal(t, int64(0), (*role).Expires)
+		}
+	}
+}
+
+func (suite *CacheTestSuite) TestSetRoleExpires() {
+	t := suite.T()
+	err := suite.cache.SetRoleExpires(TEST_ROLE_ARN, 12344553243)
+	assert.NoError(t, err)
+
+	flat, err := suite.cache.GetRole(TEST_ROLE_ARN)
+	assert.NoError(t, err)
+	assert.Equal(t, int64(12344553243), flat.ExpiresEpoch)
+
+	err = suite.cache.SetRoleExpires(INVALID_ROLE_ARN, 12344553243)
+	assert.Error(t, err)
+}
