+++ internal/sso/cache_test.go
@@ -19,17 +19,11 @@ package sso
  */
 
 import (
-	"fmt"
-	"log/slog"
 	"os"
 	"testing"
-	"time"
-
-	// "github.com/davecgh/go-spew/spew"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
-	testlogger "github.com/synfinatic/flexlog/test"
 )
 
 const (
@@ -92,260 +86,6 @@ func (suite *CacheTestSuite) TestNeedsRefresh() {
 	assert.True(t, suite.cache.SSO["Default"].NeedsRefresh(suite.settings.SSO["Default"], suite.settings))
 }
 
-func (suite *CacheTestSuite) TestAddHistory() {
-	t := suite.T()
-
-	c := &Cache{
-		settings: &Settings{
-			HistoryLimit:   1,
-			HistoryMinutes: 90,
-		},
-		ssoName: "Default",
-		SSO: map[string]*SSOCache{
-			"Default": {
-				name:       "Default",
-				LastUpdate: 2345,
-				History:    []string{},
-				Roles: &Roles{
-					Accounts: map[int64]*AWSAccount{
-						123456789012: {
-							Alias: "MyAccount",
-							Roles: map[string]*AWSRole{
-								"Foo": {
-									Arn: "arn:aws:iam::123456789012:role/Foo",
-									Tags: map[string]string{
-										"AccountAlias": "MyAccount",
-										"RoleName":     "Foo",
-									},
-								},
-								"Bar": {
-									Arn: "arn:aws:iam::123456789012:role/Bar",
-									Tags: map[string]string{
-										"AccountAlias": "MyAccount",
-										"RoleName":     "Bar",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	cache := c.GetSSO()
-	assert.Equal(t, []string{}, cache.History)
-
-	now := time.Now().Unix()
-
-	// Basic add
-	c.AddHistory("arn:aws:iam::123456789012:role/Foo")
-	assert.Equal(t, []string{"arn:aws:iam::123456789012:role/Foo"}, cache.History)
-	tag := fmt.Sprintf("MyAccount:Foo,%d", now)
-	assert.Equal(t, tag, c.GetSSO().Roles.Accounts[123456789012].Roles["Foo"].Tags["History"])
-
-	// Add again which should be a no-op
-	c.AddHistory("arn:aws:iam::123456789012:role/Foo")
-	assert.Equal(t, []string{"arn:aws:iam::123456789012:role/Foo"}, cache.History)
-	tag = fmt.Sprintf("MyAccount:Foo,%d", now)
-	assert.Equal(t, tag, c.GetSSO().Roles.Accounts[123456789012].Roles["Foo"].Tags["History"])
-
-	// Add a new item which expires the previous item
-	c.AddHistory("arn:aws:iam::123456789012:role/Bar")
-	assert.Equal(t, []string{"arn:aws:iam::123456789012:role/Bar"}, cache.History)
-	tag = fmt.Sprintf("MyAccount:Bar,%d", now)
-	assert.NotContains(t, "History", c.GetSSO().Roles.Accounts[123456789012].Roles["Foo"].Tags)
-	assert.Equal(t, tag, c.GetSSO().Roles.Accounts[123456789012].Roles["Bar"].Tags["History"])
-
-	// Add the same item again
-	c.AddHistory("arn:aws:iam::123456789012:role/Bar")
-	assert.Equal(t, []string{"arn:aws:iam::123456789012:role/Bar"}, cache.History)
-
-	// Basic tests with two items in the History slice
-	c.settings.HistoryLimit = 2
-	c.AddHistory("arn:aws:iam::123456789012:role/Foo")
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Foo",
-		"arn:aws:iam::123456789012:role/Bar"}, cache.History)
-
-	// this should be a no-op
-	c.AddHistory("arn:aws:iam::123456789012:role/Foo")
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Foo",
-		"arn:aws:iam::123456789012:role/Bar"}, cache.History)
-
-	// reorder args
-	c.AddHistory("arn:aws:iam::123456789012:role/Baz")
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Baz",
-		"arn:aws:iam::123456789012:role/Foo"}, cache.History)
-
-	c.AddHistory("arn:aws:iam::123456789012:role/Foo")
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Foo",
-		"arn:aws:iam::123456789012:role/Baz"}, cache.History)
-
-	assert.Contains(t, c.GetSSO().Roles.Accounts[123456789012].Roles["Foo"].Tags, "History")
-}
-
-func (suite *CacheTestSuite) setupDeleteOldHistory() *Cache {
-	c := &Cache{
-		settings: &Settings{
-			HistoryLimit:   2,
-			HistoryMinutes: 5,
-		},
-		ssoName: "Default",
-		SSO:     map[string]*SSOCache{},
-	}
-	now := time.Now().Unix()
-	c.SSO["Default"] = &SSOCache{
-		LastUpdate: now - 5,
-		History: []string{
-			"arn:aws:iam::123456789012:role/Test",
-			"arn:aws:iam::123456789012:role/Foo",
-		},
-		Roles: &Roles{
-			Accounts: map[int64]*AWSAccount{
-				123456789012: {
-					Roles: map[string]*AWSRole{
-						"Test": {
-							Tags: map[string]string{
-								"History": fmt.Sprintf("MyAlias:Test,%d", now-5),
-							},
-						},
-						"Foo": {
-							Tags: map[string]string{
-								"History": fmt.Sprintf("MyOtherAlias:Foo,%d", now-85),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-	return c
-}
-
-func (suite *CacheTestSuite) TestDeleteOldHistory() {
-	t := suite.T()
-
-	c := suite.setupDeleteOldHistory()
-
-	// check setup
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Test",
-		"arn:aws:iam::123456789012:role/Foo",
-	}, c.GetSSO().History)
-
-	// no-op because we haven't timed out yet
-	c.deleteOldHistory()
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Test",
-		"arn:aws:iam::123456789012:role/Foo",
-	}, c.GetSSO().History)
-
-	c = suite.setupDeleteOldHistory()
-
-	// no-op when HistoryMinutes <= 0
-	c.settings.HistoryLimit = 1
-	c.settings.HistoryMinutes = 0
-	c.deleteOldHistory()
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Test",
-		"arn:aws:iam::123456789012:role/Foo",
-	}, c.GetSSO().History)
-
-	// remove one due to HistoryLimit
-	c.settings.HistoryMinutes = 1
-	c.deleteOldHistory()
-	assert.Equal(t, []string{
-		"arn:aws:iam::123456789012:role/Test",
-	}, c.GetSSO().History)
-	assert.NotContains(t, "History",
-		c.SSO["Default"].Roles.Accounts[123456789012].Roles["Foo"].Tags)
-
-	// setup logger for tests
-	oldLogger := log.Copy()
-	tLogger := testlogger.NewTestLogger("DEBUG")
-	defer tLogger.Close()
-	log = tLogger
-
-	defer func() { log = oldLogger }()
-
-	// remove one because of HistoryMinutes expires
-	c = suite.setupDeleteOldHistory()
-	c.settings.HistoryMinutes = 1
-	c.deleteOldHistory()
-
-	msg := testlogger.LogMessage{}
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.NotEmpty(t, msg.Message)
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, "Removed expired history role", msg.Message)
-	assert.Equal(t, []string{"arn:aws:iam::123456789012:role/Test"}, c.GetSSO().History)
-
-	tLogger.Reset()
-	c = suite.setupDeleteOldHistory()
-	c.GetSSO().History = append(c.GetSSO().History, "arn:aws:iam:")
-	c.deleteOldHistory()
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, "Unable to parse History ARN", msg.Message)
-
-	tLogger.Reset()
-	c = suite.setupDeleteOldHistory()
-	c.GetSSO().History = append(c.GetSSO().History, "arn:aws:iam::123456789012:role/NoHistoryTag")
-	c.deleteOldHistory()
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, msg.Message, "but no role by that name")
-
-	tLogger.Reset()
-	c = suite.setupDeleteOldHistory()
-	c.GetSSO().History = append(c.GetSSO().History, "arn:aws:iam::1234567890:role/NoHistoryTag")
-	c.deleteOldHistory()
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, msg.Message, "but no account by that name")
-
-	tLogger.Reset()
-	c = suite.setupDeleteOldHistory()
-	c.GetSSO().History = append(c.GetSSO().History, "arn:aws:iam::123456789012:role/NoHistoryTag")
-	c.GetSSO().Roles.Accounts[123456789012].Roles["NoHistoryTag"] = &AWSRole{}
-	c.deleteOldHistory()
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, msg.Message, "in history list without a History tag")
-
-	tLogger.Reset()
-	c = suite.setupDeleteOldHistory()
-	c.GetSSO().History = append(c.GetSSO().History, "arn:aws:iam::123456789012:role/MissingHistoryTag")
-	c.GetSSO().Roles.Accounts[123456789012].Roles["MissingHistoryTag"] = &AWSRole{
-		Tags: map[string]string{
-			"History": "What:Foo",
-		},
-	}
-	c.deleteOldHistory()
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, msg.Message, "Too few fields for")
-
-	tLogger.Reset()
-	c = suite.setupDeleteOldHistory()
-	c.GetSSO().History = append(c.GetSSO().History, "arn:aws:iam::123456789012:role/MissingHistoryTag")
-	c.GetSSO().Roles.Accounts[123456789012].Roles["MissingHistoryTag"] = &AWSRole{
-		Tags: map[string]string{
-			"History": "What:Foo,kkkk",
-		},
-	}
-	c.deleteOldHistory()
-	assert.NoError(t, tLogger.GetNext(&msg))
-	assert.Equal(t, slog.LevelDebug, msg.Level)
-	assert.Contains(t, msg.Message, "Unable to parse")
-
-	tLogger.Reset()
-}
-
 func (suite *CacheTestSuite) TestExpired() {
 	t := suite.T()
 	s := SSOConfig{
@@ -366,109 +106,6 @@ func (suite *CacheTestSuite) TestExpired() {
 
 	s.settings.CacheRefresh = 1
 	assert.Error(t, suite.cache.Expired(&s))
-}
-
-func (suite *CacheTestSuite) TestGetRole() {
-	t := suite.T()
-	r, _ := suite.cache.GetRole(TEST_ROLE_ARN)
-	assert.Equal(t, int64(707513610766), r.AccountId)
-	assert.Equal(t, "AWSAdministratorAccess", r.RoleName)
-	assert.Equal(t, TEST_ROLE_ARN, r.Arn)
-	assert.Equal(t, "Dev Account", r.AccountName)
-	assert.Equal(t, "control-tower-dev-sub1-aws", r.AccountAlias)
-	assert.Equal(t, "test+control-tower-sub1@ourcompany.com", r.EmailAddress)
-	assert.Equal(t, "", r.Profile)
-	assert.Equal(t, "", r.DefaultRegion)
-	assert.Equal(t, "us-east-1", r.SSORegion)
-	assert.Equal(t, "https://d-754545454.awsapps.com/start", r.StartUrl)
-	assert.Equal(t, "Default", r.SSO)
-
-	tags := map[string]string{
-		"AccountAlias": "control-tower-dev-sub1-aws",
-		"AccountID":    "707513610766",
-		"AccountName":  "Dev Account",
-		"Email":        "test+control-tower-sub1@ourcompany.com",
-		"Role":         "AWSAdministratorAccess",
-		"Type":         "Sub Account",
-	}
-	assert.Equal(t, tags, r.Tags)
-
-	_, err := suite.cache.GetRole("arn:aws:2344:role/missing-colon")
-	assert.Error(t, err)
-}
-
-func (suite *CacheTestSuite) TestAccountIds() {
-	t := suite.T()
-
-	ids := suite.cache.GetSSO().Roles.AccountIds()
-	assert.Equal(t, 4, len(ids))
-}
-
-func (suite *CacheTestSuite) TestGetAllRoles() {
-	t := suite.T()
-
-	cache := suite.cache.GetSSO()
-	roles := cache.Roles.GetAllRoles()
-	assert.Equal(t, 19, len(roles))
-
-	aroles := cache.Roles.GetAccountRoles(707513610766)
-	assert.Equal(t, 4, len(aroles))
-	aroles = cache.Roles.GetAccountRoles(502470824893)
-	assert.Equal(t, 4, len(aroles))
-	aroles = cache.Roles.GetAccountRoles(25823461518)
-	assert.Equal(t, 7, len(aroles))
-}
-
-func (suite *CacheTestSuite) TestGetAllTags() {
-	t := suite.T()
-	cache := suite.cache.GetSSO()
-
-	tl := cache.Roles.GetAllTags()
-	assert.Equal(t, 10, len(*tl))
-	tl = suite.cache.GetAllTagsSelect()
-	assert.Equal(t, 10, len(*tl))
-}
-
-func (suite *CacheTestSuite) TestGetRoleTags() {
-	t := suite.T()
-	cache := suite.cache.GetSSO()
-
-	rt := cache.Roles.GetRoleTags()
-	assert.Equal(t, 19, len(*rt))
-	rt = suite.cache.GetRoleTagsSelect()
-	assert.Equal(t, 19, len(*rt))
-}
-
-func (suite *CacheTestSuite) TestMatchingRoles() {
-	t := suite.T()
-	cache := suite.cache.GetSSO()
-
-	match := map[string]string{
-		"Type": "Main Account",
-		"Foo":  "Bar",
-	}
-	roles := cache.Roles.MatchingRoles(match)
-	assert.Equal(t, 1, len(roles))
-
-	match["Nothing"] = "Matches"
-	roles = cache.Roles.MatchingRoles(match)
-	assert.Equal(t, 0, len(roles))
-}
-
-func (suite *CacheTestSuite) TestIsExpired() {
-	t := suite.T()
-
-	r, _ := suite.cache.GetRole(TEST_ROLE_ARN)
-	assert.True(t, r.IsExpired())
-}
-
-func (suite *CacheTestSuite) BadRole() {
-	t := suite.T()
-	_, err := suite.cache.GetRole(INVALID_ROLE_ARN)
-	assert.Error(t, err)
-
-	_, err = suite.cache.GetRole(INVALID_ACCOUNT_ARN)
-	assert.Error(t, err)
 }
 
 func (suite *CacheTestSuite) TestVersion() {
@@ -499,76 +136,6 @@ func (suite *CacheTestSuite) TestGetSSO() {
 func (suite *CacheTestSuite) TestCacheFile() {
 	t := suite.T()
 	assert.Equal(t, suite.cacheFile, suite.cache.CacheFile())
-}
-
-func (suite *CacheTestSuite) TestMarkRolesExpired() {
-	t := suite.T()
-	err := suite.cache.MarkRolesExpired()
-	assert.NoError(t, err)
-
-	sso := suite.cache.GetSSO()
-	for _, account := range sso.Roles.Accounts {
-		for _, role := range account.Roles {
-			assert.Equal(t, int64(0), (*role).Expires)
-		}
-	}
-}
-
-func (suite *CacheTestSuite) TestSetRoleExpires() {
-	t := suite.T()
-	err := suite.cache.SetRoleExpires(TEST_ROLE_ARN, 12344553243)
-	assert.NoError(t, err)
-
-	flat, err := suite.cache.GetRole(TEST_ROLE_ARN)
-	assert.NoError(t, err)
-	assert.Equal(t, int64(12344553243), flat.ExpiresEpoch)
-
-	err = suite.cache.SetRoleExpires(INVALID_ROLE_ARN, 12344553243)
-	assert.Error(t, err)
-}
-
-func (suite *CacheTestSuite) TestProcessSSORoles() {
-	t := suite.T()
-
-	roles := []RoleInfo{
-		{
-			Id:           0,
-			Arn:          "arn:aws:iam::123456789012:role/testing",
-			RoleName:     "testing",
-			AccountId:    "123456789012",
-			AccountName:  "MyTestAccount",
-			EmailAddress: "testing@domain.com",
-			Expires:      555555555,
-			Profile:      "TestingAccountTesting",
-			Region:       "us-east-1",
-			SSORegion:    "us-east-2",
-			StartUrl:     "https://fake.awsapps.com/start",
-		},
-	}
-
-	r := Roles{
-		Accounts: map[int64]*AWSAccount{},
-	}
-	cache := SSOCache{
-		Roles: &Roles{
-			Accounts: map[int64]*AWSAccount{
-				123456789012: {},
-			},
-		},
-	}
-
-	processSSORoles(roles, &cache, &r)
-	assert.Len(t, r.Accounts, 1)
-	assert.Len(t, r.Accounts[123456789012].Roles, 1)
-	assert.Equal(t, "MyTestAccount", r.Accounts[123456789012].Alias)
-	assert.Equal(t, "testing@domain.com", r.Accounts[123456789012].EmailAddress)
-	assert.Equal(t, "arn:aws:iam::123456789012:role/testing", r.Accounts[123456789012].Roles["testing"].Arn)
-	assert.Len(t, r.Accounts[123456789012].Tags, 0)
-
-	assert.Equal(t, "MyTestAccount", r.Accounts[123456789012].Roles["testing"].Tags["AccountAlias"])
-	assert.Equal(t, "123456789012", r.Accounts[123456789012].Roles["testing"].Tags["AccountID"])
-	assert.Equal(t, "testing@domain.com", r.Accounts[123456789012].Roles["testing"].Tags["Email"])
-	assert.Equal(t, "testing", r.Accounts[123456789012].Roles["testing"].Tags["Role"])
 }
 
 func (suite *CacheTestSuite) TestPruneSSO() {
@@ -623,67 +190,6 @@ func (suite *CacheTestSuite) TestGetSSOByName() {
 	assert.NotNil(t, cache)
 	assert.Empty(t, cache.name)
 }
-
-func (suite *CacheTestSuite) TestAddConfigRoles() {
-	t := suite.T()
-
-	config := &SSOConfig{
-		Accounts: map[string]*SSOAccount{
-			"123456789012": {
-				Roles: map[string]*SSORole{
-					"Foo": {},
-					"Bar": {},
-				},
-			},
-		},
-	}
-
-	roles := &Roles{
-		Accounts: map[int64]*AWSAccount{},
-	}
-
-	err := suite.cache.addConfigRoles(roles, config)
-	assert.NoError(t, err)
-}
-
-/*
-func (suite *CacheTestSuite) TestCacheRefresh() {
-	t := suite.T()
-
-	settings := &Settings{
-		SSO: map[string]*SSOConfig{
-			"Default": {
-				SSORegion:     "us-east-1",
-				DefaultRegion: "us-east-1",
-			},
-		},
-		HistoryLimit:   1,
-		HistoryMinutes: 90,
-		DefaultSSO:     "Default",
-		cacheFile:      suite.cacheFile,
-		ProfileFormat:  "{{ .AccountIdPad }}:{{ .RoleName }}",
-	}
-	settings.SSO["Default"].settings = settings
-
-	storeFile, err := os.CreateTemp("", "*")
-	assert.NoError(t, err)
-	storeFile.Close()
-	defer os.Remove(storeFile.Name())
-
-	jstore, err := storage.OpenJsonStore(storeFile.Name())
-	assert.NoError(t, err)
-
-	sso := NewAWSSSO(suite.settings.SSO["Default"], &jstore)
-
-	// Ensure the SecureStorage interface is implemented correctly
-	added, deleted, err := suite.cache.Refresh(sso, settings.SSO["Default"], "Default", 1)
-	assert.NoError(t, err)
-	assert.Equal(t, int64(0), suite.cache.ConfigCreatedAt)
-	assert.Equal(t, int64(1), suite.cache.Version)
-	assert.Equal(t, 1, added)
-	assert.Equal(t, 0, deleted)
-}
-*/
 
 func TestOpenCacheFailure(t *testing.T) {
 	s := &Settings{}
