+++ internal/sso/settings.go
@@ -40,10 +40,6 @@ import (
 	"github.com/synfinatic/aws-sso-cli/internal/uri"
 )
 
-const (
-	NIX_STORE_PREFIX = "/nix/store/"
-)
-
 type Settings struct {
 	configFile                string                   // name of this file
 	cacheFile                 string                   // name of cache file; always passed in via CLI args
@@ -404,141 +400,4 @@ func (s *Settings) GetEnvVarTags() map[string]string {
 		ret[tag] = fmt.Sprintf("AWS_SSO_TAG_%s", strings.ToUpper(tag))
 	}
 	return ret
-}
-
-type ProfileMap map[string]map[string]ProfileConfig
-
-type ProfileConfig struct {
-	Arn             string
-	BinaryPath      string
-	ConfigVariables map[string]interface{}
-	DefaultRegion   string
-	Open            string
-	Profile         string
-	Sso             string
-}
-
-// allow os.Executable call to be overridden for unit testing purposes
-var getExecutable func() (string, error) = func() (string, error) {
-	exec, err := os.Executable()
-	if err != nil {
-		return "", err
-	}
-	if strings.HasPrefix(exec, NIX_STORE_PREFIX) {
-		log.Warn("Detected NIX. Using $PATH to find `aws-sso`. Override with `ConfigProfilesBinaryPath`")
-		exec = "aws-sso"
-	}
-	return exec, nil
-}
-
-// GetAllProfiles returns a map of the ProfileConfig for each SSOConfig.
-// takes the binary path to `open` URL with if set
-func (s *Settings) GetAllProfiles() (*ProfileMap, error) {
-	profiles := ProfileMap{}
-	var err error
-
-	binaryPath := s.ConfigProfilesBinaryPath
-	if binaryPath == "" {
-		if binaryPath, err = getExecutable(); err != nil {
-			return &profiles, err
-		}
-	}
-
-	// Find all the roles across all of the SSO instances
-	for ssoName, sso := range s.Cache.SSO {
-		for _, role := range sso.Roles.GetAllRoles() {
-			profile, err := role.ProfileName(s)
-			if err != nil {
-				return &profiles, err
-			}
-
-			if _, ok := profiles[ssoName]; !ok {
-				profiles[ssoName] = map[string]ProfileConfig{}
-			}
-
-			profiles[ssoName][role.Arn] = ProfileConfig{
-				Arn:             role.Arn,
-				BinaryPath:      binaryPath,
-				ConfigVariables: s.ConfigVariables,
-				DefaultRegion:   role.DefaultRegion,
-				Profile:         profile,
-				Sso:             ssoName,
-			}
-		}
-	}
-
-	return &profiles, nil
-}
-
-// GetAllProfiles returns a map of the ProfileConfig for each SSOConfig.
-// takes the binary path to `open` URL with if set
-func (s *Settings) GetSSOProfiles(ssoName string) (*ProfileMap, error) {
-	profiles := ProfileMap{}
-	var err error
-
-	binaryPath := s.ConfigProfilesBinaryPath
-	if binaryPath == "" {
-		if binaryPath, err = getExecutable(); err != nil {
-			return &profiles, err
-		}
-	}
-
-	// Find all the roles across all of the SSO instances
-	//	for ssoName, sso := range s.Cache.SSO {
-	sso := s.Cache.GetSSOByName(ssoName)
-	for _, role := range sso.Roles.GetAllRoles() {
-		profile, err := role.ProfileName(s)
-		if err != nil {
-			return &profiles, err
-		}
-
-		if _, ok := profiles[ssoName]; !ok {
-			profiles[ssoName] = map[string]ProfileConfig{}
-		}
-
-		profiles[ssoName][role.Arn] = ProfileConfig{
-			Arn:             role.Arn,
-			BinaryPath:      binaryPath,
-			ConfigVariables: s.ConfigVariables,
-			DefaultRegion:   role.DefaultRegion,
-			Profile:         profile,
-			Sso:             ssoName,
-		}
-	}
-	// }
-
-	return &profiles, nil
-}
-
-// UniqueCheck verifies that all of the profiles are unique.
-// Only checks profiles that are actually in the ProfileMap rather than
-// iterating over all cached SSO instances, which could produce false
-// duplicates when the ProfileFormat includes {{ .SSO }}.
-func (p *ProfileMap) UniqueCheck(s *Settings) error {
-	profileUniqueCheck := map[string][]string{} // ProfileName() => Arn
-
-	for ssoName, roles := range *p {
-		for arn, config := range roles {
-			profile := config.Profile
-
-			if match, duplicate := profileUniqueCheck[profile]; duplicate {
-				return fmt.Errorf("duplicate profile name '%s' for:\n%s: %s\n%s: %s",
-					profile, match[0], match[1], ssoName, arn)
-			}
-			profileUniqueCheck[profile] = []string{ssoName, arn}
-		}
-	}
-
-	return nil
-}
-
-func (p *ProfileMap) IsDuplicate(newProfile string) bool {
-	for _, roles := range *p {
-		for _, config := range roles {
-			if config.Profile == newProfile {
-				return true
-			}
-		}
-	}
-	return false
 }
